fix(querylog): build SQLite target only on modernc-supported platforms
The pure-Go SQLite driver chain added in #2080 (github.com/glebarez/sqlite
-> modernc.org/sqlite -> modernc.org/libc) ships no generated code for
several release build targets (linux/mips, linux/mipsle, netbsd/arm,
netbsd/arm64, openbsd/arm), so `goreleaser build` failed with
"build constraints exclude all Go files in modernc.org/libc".

Move the SQLite dialector behind a build constraint so it is compiled only
on the GOOS/GOARCH combinations modernc.org/sqlite actually supports. On the
remaining targets a stub returns a clear runtime error, so those binaries
still build (use the csv, mysql or postgresql query log target there).
