+++ querylog/database_writer_test.go
@@ -1,3 +1,5 @@
+//go:build !mips && !mipsle && !mips64 && !mips64le && !(netbsd && !amd64) && !(openbsd && !amd64 && !arm64)
+
 package querylog
 
 import (
